add component Heade, check token, log out

+++ src/services/todoServices.js
@@ -1,4 +1,9 @@
+import { isValid } from "./authServices";
+
 export const getTasks= async () =>{
+  const token = await isValid()
+  if(!token) return false
+
   try{
     const res = await fetch('http://localhost:3000/todos');
     if(!res.ok){
@@ -11,6 +16,9 @@ export const getTasks= async () =>{
 }
 
 export const addTask = async(task) => {
+  const token = await isValid()
+  if(!token) return false
+  
   try{
     const res = await fetch('http://localhost:3000/todos', {
       method: 'POST',
@@ -29,6 +37,9 @@ export const addTask = async(task) => {
 }
 
 export const updateTask = async (id, updates) => {
+  const token = await isValid()
+  if(!token) return false
+
   try{
     const res = await fetch(`http://localhost:3000/todos/${id}`, {
         method: 'PATCH',
@@ -48,6 +59,9 @@ export const updateTask = async (id, updates) => {
 }
 
 export const deleteTask = async (id) => {
+  const token = await isValid()
+  if(!token) return false
+
   try{
     const res = await fetch(`http://localhost:3000/todos/${id}`, {
         method: 'DELETE'
